+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Use the encoded resource path and query parameters for Mauth singer
+- Fall back to V1 authentication when V2 authentication fails using Akka Http
 
 ## [4.0.0] - 2020-03-19
 ### Added